chore: move balance type to balance.rs and anchors to chain_data
