+++ crates/chain/src/lib.rs
@@ -23,15 +23,15 @@
 pub use bitcoin;
 mod spk_txout_index;
 pub use spk_txout_index::*;
+mod balance;
 mod chain_data;
 pub use chain_data::*;
 pub mod indexed_tx_graph;
+pub use balance::*;
 pub use indexed_tx_graph::IndexedTxGraph;
 pub mod keychain;
 pub mod local_chain;
-mod tx_data_traits;
 pub mod tx_graph;
-pub use tx_data_traits::*;
 pub use tx_graph::TxGraph;
 mod chain_oracle;
 pub use chain_oracle::*;
